⚽️ refactor: centralize background color configuration into a shared utility and improve path resolution error handling
