+++ src/themes/blur-theme/platforms/windows.js
@@ -3,8 +3,6 @@ const vscode      = require('vscode');
 const fs          = require('fs');
 const fsPromises  = require('fs').promises;
 const path        = require('path');
-const os          = require('os');
-
 const {
   generateNewJS, removeJSMarkers,
   injectElectronOptions, removeElectronOptions,
@@ -20,47 +18,10 @@ const {
 const RUNTIME_VERSION = 'v1';
 const RUNTIME_DIR_NAME = `lynx-blur-runtime-${RUNTIME_VERSION}`;
 
-// ─── Background transparency keys (same groups as Linux) ──────────────────────
-
-const TRANSPARENT_BG_KEYS = [
-  'editorPane.background',
-  'editorGroupHeader.tabsBackground',
-  'editorGroupHeader.noTabsBackground',
-  'breadcrumb.background',
-  'editorGutter.background',
-  'panel.background',
-  'panelStickyScroll.background',
-  'tab.activeBackground',
-  'tab.unfocusedActiveBackground',
-];
-
-const SEMITRANSPARENT_BG_KEYS = [
-  'sideBar.background',
-  'sideBarTitle.background',
-  'sideBarStickyScroll.background',
-  'editor.background',
-  'editorStickyScroll.background',
-  'editorStickyScrollGutter.background',
-  'tab.inactiveBackground',
-  'tab.unfocusedInactiveBackground',
-];
-
-const OPAQUE_BG_KEYS = [
-  'inlineChat.background',
-  'editorWidget.background',
-  'editorHoverWidget.background',
-  'editorSuggestWidget.background',
-  'notifications.background',
-  'notificationCenterHeader.background',
-  'menu.background',
-  'quickInput.background',
-];
-
-const ALL_BG_KEYS = [...TRANSPARENT_BG_KEYS, ...SEMITRANSPARENT_BG_KEYS, ...OPAQUE_BG_KEYS];
-
-// Lynx Dark glassmorphism base color
-const THEME_BG = '060a08';
-const DEFAULT_OPACITY = 0.45;
+const {
+  TRANSPARENT_BG_KEYS, SEMITRANSPARENT_BG_KEYS, OPAQUE_BG_KEYS,
+  ALL_BG_KEYS, THEME_BG, DEFAULT_OPACITY,
+} = require('../utils/color-keys');
 
 // ─── In-memory mutex ──────────────────────────────────────────────────────────
